build: make just ci gate again instead of dying at clippy
`just ci` was RED at clippy, and because the gate is SEQUENTIAL that meant
lint-domains, deny, check-matrix, check, doc, test and test-integration never
ran at all. The gate reported red while gating nothing, and the failure was
recorded in the handoff as "a pending Task 7/8 migration gate" rather than as a
gate outage.

Root cause: the `clippy.toml` host-authority catalog is a CENSUS input, not a
build gate. `scripts/migrate/check-host-authority-transitions.py` (via
`just lint-domains`) is the semantic authority for reviewed host-facility uses
and captures them with `--force-warn clippy::disallowed_methods`, which pierces
any ordinary level. Left at its warn default, `just clippy -- -D warnings`
promoted all ~682 catalogued product uses to hard errors. Setting
`disallowed_methods = "allow"` in `[workspace.lints.clippy]` hides nothing from
the census and stops the double gate.

That unblocked two layers that had NEVER been linted, because their crates were
dependencies of the crates failing first:

- `carrick-runtime` (13): inconsistent digit grouping in nine synthetic test
  ids, `len() != 0`, a redundant `u64` cast, a `drop()` of a non-`Drop` value,
  and a five-field `Arc<Mutex<Vec<(u64, u64, u64, u64, u64)>>>` that is now the
  named `InheritedStateRow`.
- `carrick-vmm-hvf` (10): twelve test modules were spelled
  `#[cfg(all(test, target_os = ..., target_arch = ...))]`, and clippy's
  `allow-unwrap-in-tests` does not recognize that nested form — so test code was
  being linted as production. Split into `#[cfg(test)]` plus
  `#[cfg(all(target_os = ..., target_arch = ...))]`, which is semantically
  identical since cfg attributes AND. The `initial_runner_handoff` integration
  target allows test panics explicitly, because `allow-expect-in-tests` exempts
  only `#[test]` functions and the offender is a shared fixture helper.
- `carrick-x86` (1): `decode_x86_resume_metadata` returned a five-field tuple
  read back positionally as `metadata.0 .. metadata.4`. Each member is a
  distinct domain (resume PC, raw syscall number, canonical syscall number,
  sysret frame, fork-child flag), so it is now the named `X86ResumeMetadata` —
  the same discipline `docs/typed-interfaces-audit.md` applies elsewhere.

Verified: `cargo clippy --workspace --all-targets -- -D warnings` passes, and
`just ci` now advances PAST clippy for the first time. Note for future runs:
`just ci | tail` reports `tail`'s exit status, not the gate's — redirect to a
file and read `$?`, per the gate-log rule in AGENTS.md.

+++ crates/carrick-runtime/src/dispatch/tests.rs
@@ -1058,7 +1058,7 @@ mod overlay_dispatch_tests {
             [
                 (-100i64) as u64,
                 path,
-                (O_CREAT | 2) as u64,
+                O_CREAT | 2,
                 0o600,
                 0,
                 0,
